+++ vision_agent/tools/meta_tools_types.py
